Migration to go modules (#1178)
* Migration to go modules

* Do not override version for gravitational/kingpin

In previous commit fca76d67 gravitational/kingpin version
was overriden by setting to the same version as kingpin.v2. But
gravitational/kingping is the dependency for teleport and doesn't used
in any of binaries in this repo. Reverted version to the one from
teleport repository.

* Support Go modules in Makefiles

* Change version of Golang to 1.13.8 for better support of modules.
* Change ldflags for gravitational/version linking. With Go modules
path to the dependencies is relative to `vendor` directory.
* Delete unused code for godep.
* Update go build/install commands to force them use of Go modules.
* Pass current user/group as a build argument to buildbox container.
It will help to avoid issues with permissions and in future we can
mount cache directory from host to speed up builds.